using controllers  to refactor the routers code
